Rename files to lower case.
