+++ packages/full-text-search/src/fuzzy/automaton.ts
