+++ packages/full-text-search/src/fuzzy/long.ts
